requirements.txt
ultralytics==8.3.0
fastapi
uvicorn
pillow
torch
python-multipart